+++ src/components/form/b-select/test/unit/simple.ts
@@ -8,6 +8,7 @@
 
 import type { Page } from 'playwright';
 
+import { Component } from 'tests/helpers';
 import test from 'tests/config/unit/test';
 
 import { assertValueIs, createSelector, renderSelect, selectValue } from 'components/form/b-select/test/helpers';
@@ -185,6 +186,31 @@ test.describe('<b-select> simple usage', () => {
 				['SPAN', 'Bar']
 			]);
 		});
+
+		test('should close the dropdown when click on an element with .stopPropagation', async ({page}) => {
+			const btnText = 'buttonWithStopPropagation';
+			await Component.createComponent(page, 'b-button', {
+				attrs: {
+					'@click.stop': () => {},
+				},
+				children: {
+					default: btnText,
+				}
+			});
+
+			const target = await renderSelect(page, {
+				items: [
+					{label: 'Foo', value: 0},
+					{label: 'Bar', value: 1}
+				]
+			});
+
+			await target.evaluate(async ctx => await ctx.open());
+			await test.expect(page.locator(createSelector('dropdown')).isVisible()).resolves.toBeTruthy();
+
+			await page.getByText(btnText).click();
+			await test.expect(page.locator(createSelector('dropdown')).isHidden()).resolves.toBeTruthy();
+		})
 
 		test('should be rendered to a native <select> with `native = true`', async ({page}) => {
 			const target = await renderSelect(page, {