+++ src/components/traits/i-open/i-open.ts
@@ -74,7 +74,7 @@ export default abstract class iOpen {
 	 * @param component
 	 * @param [events] - a map with events to listen
 	 */
-	static initCloseHelpers<T extends iBlock>(component: T & iOpen, events: CloseHelperEvents = {}): void {
+	static initCloseHelpers<T extends iBlock>(component: T & iOpen, events: CloseHelperEvents = {}, eventOpts: AddEventListenerOptions = {}): void {
 		const {
 			async: $a,
 			localEmitter: $e
@@ -92,7 +92,10 @@ export default abstract class iOpen {
 			$a.setTimeout(() => {
 				const opts = {
 					...helpersGroup,
-					options: {passive: false}
+					options: {
+						passive: false,
+						...eventOpts,
+					}
 				};
 
 				try {